PlaylistPorting/requirements.txt
Unidecode==0.04.21
discogs_client==2.2.1
gmusicapi==10.1.2
pocket_api==0.1.5
pylast==1.9.0
pandas==0.19.2
youtube_dl==2017.7.30.1
spotipy==2.4.4
todoist_python==7.0.17
google_auth_oauthlib==0.1.1
requests==2.18.4
eyeD3==0.8
PyLyrics==1.1.0
google_api_python_client==1.6.4
billboard.py==4.0.0
beautifulsoup4==4.6.0
pocket==0.3.6
todoist==0.0.1
